I made a more proper draw() abstraction for DGImage.  It's still not totally what Tom wants, but if nothing else, it makes the current controls' draw methods a little saner (and less redundant).
I also moved over to using signed long rather than unsigned long as
the returns for DGImage's getHeight and getWidth methods and also as
the port height argument for the various DGRect -> CGRect and draw
methods.  This fixed a few bugs that it turns out were happening
during some arithmatic where the entire operations were quietly being
cast to unsigned long rather than signed long, and wraparound was
occurring (for example, that's what was causing the weird clipping in
AudioUnitHosting when you opened an AU that had a UI view that was
taller than the current size of AUH's window).

I also trid experimenting with transforming the CoreGraphics drawing
matrix so that I could hopefully use non-upside-down coordinates.  I
was able to this and successfully position things with non-upside-down
coordinates, but the big drawback was that everything that got drawn
(images and text) was drawn upside down.  Hmmm...

+++ dfxgui/dfxguibutton.h
@@ -27,7 +27,7 @@ public:
 				long inNumStates, DfxGuiBottonMode inMode, bool inDrawMomentaryState = false);
 	virtual ~DGButton();
 
-	virtual void draw(CGContextRef context, UInt32 portHeight);
+	virtual void draw(CGContextRef context, long portHeight);
 	virtual void mouseDown(float inXpos, float inYpos, unsigned long inMouseButtons, unsigned long inKeyModifiers);
 	virtual void mouseTrack(float inXpos, float inYpos, unsigned long inMouseButtons, unsigned long inKeyModifiers);
 	virtual void mouseUp(float inXpos, float inYpos, unsigned long inKeyModifiers);
